chore: enable typed linting for IDE
