+++ package.json
@@ -212,8 +212,8 @@
     "link.dist.npm": "cd packages/qwik && npm link && cd ../qwik-router && npm link && cd ../eslint-plugin-qwik && npm link && cd ../qwik-react && npm link",
     "link.dist.yarn": "cd packages/qwik && yarn link && cd ../qwik-router && yarn link && cd ../eslint-plugin-qwik && yarn link && cd ../qwik-react && yarn link",
     "lint": "pnpm lint.eslint && pnpm lint.prettier && pnpm lint.rust",
-    "lint.eslint": "eslint --cache \"**/*.ts*\" && pnpm -r --parallel lint",
-    "lint.fix": "eslint --fix \"**/*.ts*\" && pnpm -r --parallel lint.fix && pnpm prettier.fix",
+    "lint.eslint": "eslint -c eslint.config-no-ts.mjs && pnpm -r --parallel lint",
+    "lint.fix": "eslint --fix -c eslint.config-no-ts.mjs && pnpm -r --parallel lint.fix && pnpm prettier.fix",
     "lint.prettier": "prettier --cache --check .",
     "lint.rust": "make lint",
     "lint.syncpack": "syncpack list-mismatches",
